pyproject.toml
[tool.poetry]
name = "movies-api"
version = "0.1.0"
description = ""
authors = ["Soraia Tavares <soraiaoliveira094@email.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.68.1"
uvicorn = "^0.15.0"
httpx = "^0.19.0"
asyncpg = "^0.24.0"
psycopg2-binary = "^2.9.3"
strawberry-graphql = "^0.83.1"

[tool.poetry.dev-dependencies]
black = "^21.9b0"
flake8 = "^5.0.1"
isort = "^5.9.3"
